requirements.txt
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==3.3.1
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==8.0.1
cloudpickle==1.6.0
dask==2021.7.1
ddtrace==0.44.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
dnspython==2.1.0
docker==4.3.1
entrypoints==0.3
filelock==3.0.12
fire==0.3.1
Flask==2.0.1
fsspec==2021.7.0
graphviz==0.14.1
grpcio==1.37.1
grpcio-tools==1.37.1
gunicorn==20.0.4
huggingface-hub==0.0.12
idna==2.10
intervaltree==3.1.0
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
locket==0.2.1
loguru==0.5.3
lz4==3.1.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
mypy-protobuf==1.23
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5
nltk==3.6.2
notebook==6.4.0
packaging==20.9
pandas==1.2.5
pandocfilters==1.4.3
parso==0.8.2
partd==1.2.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pinecone-client==0.8.59
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.13.0
psutil==5.8.0
ptyprocess==0.7.0
pycparser==2.20
pydot==1.4.1
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
PyYAML==5.4
pyzmq==19.0.2
regex==2021.4.4
requests==2.25.1
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.0
Send2Trash==1.7.1
sentence-transformers==2.0.0
sentencepiece==0.1.96
sentry-sdk==1.0.0
six==1.16.0
sortedcontainers==2.4.0
swifter==1.0.9
tenacity==7.0.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.1.0
tokenizers==0.10.3
toolz==0.11.1
torch==1.9.0
torchvision==0.10.0
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
transformers==4.8.1
typing-extensions==3.7.4.3
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.1
widgetsnbextension==3.5.1
